email sending for low stocks dashboard fixed
	modified:   README.md
	modified:   app/Filament/Resources/OrderResource.php
	modified:   app/Filament/Resources/OrderResource/Pages/CreateOrder.php
	modified:   app/Filament/Resources/OrderResource/Pages/ViewOrder.php
	modified:   app/Filament/Resources/ProductCategoryResource.php
	new file:   app/Http/Controllers/SendMailController.php
	new file:   app/Mail/SampleMail.php
	modified:   app/Providers/Filament/AdminPanelProvider.php
	modified:   composer.json
	modified:   composer.lock
	modified:   database/migrations/2024_02_19_085020_create_orders_table.php
	modified:   deploy.sh
	new file:   public/css/swisnl/filament-backgrounds/filament-backgrounds-styles.css
	new file:   public/images/swisnl/filament-backgrounds/curated-by-swis/01.jpg
	new file:   public/images/swisnl/filament-backgrounds/curated-by-swis/02.jpg
	new file:   public/images/swisnl/filament-backgrounds/curated-by-swis/03.jpg
	new file:   public/images/swisnl/filament-backgrounds/curated-by-swis/04.jpg
	new file:   public/images/swisnl/filament-backgrounds/curated-by-swis/05.jpg
	new file:   public/images/swisnl/filament-backgrounds/curated-by-swis/06.jpg
	new file:   public/images/swisnl/filament-backgrounds/curated-by-swis/07.jpg
	new file:   public/images/swisnl/filament-backgrounds/curated-by-swis/08.jpg
	new file:   public/images/swisnl/filament-backgrounds/curated-by-swis/09.jpg
	new file:   public/images/swisnl/filament-backgrounds/curated-by-swis/10.jpg
	new file:   public/images/swisnl/filament-backgrounds/curated-by-swis/11.jpg
	new file:   public/images/swisnl/filament-backgrounds/curated-by-swis/12.jpg
	new file:   public/images/swisnl/filament-backgrounds/curated-by-swis/13.jpg
	new file:   public/images/swisnl/filament-backgrounds/curated-by-swis/14.jpg
	new file:   public/images/swisnl/filament-backgrounds/curated-by-swis/15.jpg
	new file:   public/images/swisnl/filament-backgrounds/curated-by-swis/16.jpg
	new file:   public/images/swisnl/filament-backgrounds/curated-by-swis/17.jpg
	new file:   public/images/swisnl/filament-backgrounds/curated-by-swis/18.jpg
	new file:   public/images/swisnl/filament-backgrounds/curated-by-swis/19.jpg
	new file:   public/images/swisnl/filament-backgrounds/curated-by-swis/20.jpg
	new file:   public/images/swisnl/filament-backgrounds/curated-by-swis/21.jpg
	new file:   public/images/swisnl/filament-backgrounds/curated-by-swis/22.jpg
	new file:   public/images/swisnl/filament-backgrounds/curated-by-swis/23.jpg
	new file:   public/images/swisnl/filament-backgrounds/curated-by-swis/24.jpg
	new file:   public/images/swisnl/filament-backgrounds/curated-by-swis/25.jpg
	new file:   public/images/swisnl/filament-backgrounds/curated-by-swis/26.jpg
	new file:   public/images/swisnl/filament-backgrounds/curated-by-swis/27.jpg
	new file:   public/images/swisnl/filament-backgrounds/curated-by-swis/28.jpg
	new file:   public/images/swisnl/filament-backgrounds/curated-by-swis/29.jpg
	new file:   public/images/swisnl/filament-backgrounds/curated-by-swis/30.jpg
	new file:   public/images/swisnl/filament-backgrounds/curated-by-swis/31.jpg
	new file:   public/images/swisnl/filament-backgrounds/triangles/01.svg
	new file:   public/images/swisnl/filament-backgrounds/triangles/02.svg
	new file:   public/images/swisnl/filament-backgrounds/triangles/03.svg
	new file:   public/images/swisnl/filament-backgrounds/triangles/04.svg
	new file:   public/images/swisnl/filament-backgrounds/triangles/05.svg
	new file:   public/images/swisnl/filament-backgrounds/triangles/06.svg
	new file:   public/images/swisnl/filament-backgrounds/triangles/07.svg
	new file:   public/images/swisnl/filament-backgrounds/triangles/08.svg
	new file:   public/images/swisnl/filament-backgrounds/triangles/09.svg
	new file:   public/images/swisnl/filament-backgrounds/triangles/10.svg
	new file:   public/images/swisnl/filament-backgrounds/triangles/11.svg
	new file:   public/images/swisnl/filament-backgrounds/triangles/12.svg
	new file:   public/images/swisnl/filament-backgrounds/triangles/13.svg
	new file:   public/images/swisnl/filament-backgrounds/triangles/14.svg
	new file:   public/images/swisnl/filament-backgrounds/triangles/15.svg
	new file:   public/images/swisnl/filament-backgrounds/triangles/16.svg
	new file:   public/images/swisnl/filament-backgrounds/triangles/17.svg
	new file:   public/images/swisnl/filament-backgrounds/triangles/18.svg
	new file:   public/images/swisnl/filament-backgrounds/triangles/19.svg
	new file:   public/images/swisnl/filament-backgrounds/triangles/20.svg
	new file:   public/images/swisnl/filament-backgrounds/triangles/21.svg
	new file:   public/images/swisnl/filament-backgrounds/triangles/22.svg
	new file:   public/images/swisnl/filament-backgrounds/triangles/23.svg
	new file:   public/images/swisnl/filament-backgrounds/triangles/24.svg
	new file:   public/images/swisnl/filament-backgrounds/triangles/25.svg
	new file:   resources/views/emails/low-stocks.blade.php
	modified:   routes/web.php

+++ database/migrations/2024_02_19_085020_create_orders_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->string("client_name");
             $table->string("client_phone");
             $table->string("client_address");
-            $table->decimal('total', 8, 2);
+            $table->double('total', 100, 2);
             $table->boolean('delivered')->default(false);
             $table->timestamps();
             $table->softDeletes();
